requirements.txt
python-telegram-bot==20.7
python-binance==1.0.15
pandas==2.2.3
ta==0.11.2
python-dotenv==1.0.0
requests==2.31.0
